JS update
Selected h1 to add title to DOM.

Removed all pokemonvobjects inside pokemonList array.

Added 'event' as a argument inside the addListItem function in the addEventListener.

added LoadList as a function to load the list of the pokemon and added loadDetails to provide specific details for each pokemon.

added loadList, loadDetails, showDetails im the return function.